布隆过滤器有bug，暂时没找到原因

+++ RedisDemo.Redis支持的存储数据类型/布隆过滤器/BloomFilterBase.cs
@@ -31,7 +31,7 @@ namespace RedisDemo
         /// </summary>
         /// <param name="capacity">The anticipated number of items to be added to the filter. More than this number of items can be added, but the error rate will exceed what is expected.</param>
         /// <param name="errorRate">The accepable false-positive rate (e.g., 0.01F = 1%)</param>
-        public BloomFilterBase(int capacity, int errorRate) : this(capacity, errorRate, null) { }
+        public BloomFilterBase(int capacity, float errorRate) : this(capacity, errorRate, null) { }
 
         /// <summary>
         /// Creates a new Bloom filter, specifying an error rate of 1/capacity, using the optimal size for the underlying data structure based on the desired capacity and error rate, as well as the optimal number of hash functions.
